move generic Feature tests to dedicated source file
