+++ src/GeoJSON.Net.Tests/Feature/FeatureTests.cs
@@ -30,31 +30,6 @@ namespace GeoJSON.Net.Tests.Feature
             Assert.AreEqual(feature.Geometry.Type, GeoJSONObjectType.Point);
         }
 
-        private class TypedFeatureProps
-        {
-            [JsonProperty("name")]
-            public string Name { get; set; }
-            [JsonProperty("value")]
-            public double Value { get; set; }
-        }
-
-        [Test]
-        public void Can_Deserialize_Typed_Point_Feature()
-        {
-            var json = GetExpectedJson();
-            var feature = JsonConvert.DeserializeObject<Feature<Point, TypedFeatureProps>>(json);
-
-            Assert.IsNotNull(feature);
-
-            Assert.IsNotNull(feature.Properties);
-            Assert.AreEqual(feature.Properties.Name, "Dinagat Islands");
-            Assert.AreEqual(feature.Properties.Value, 4.2);            
-
-            Assert.AreEqual(feature.Id, "test-id");
-
-            Assert.AreEqual(feature.Geometry.Type, GeoJSONObjectType.Point);
-        }
-
         [Test]
         public void Can_Serialize_LineString_Feature()
         {
@@ -123,23 +98,6 @@ namespace GeoJSON.Net.Tests.Feature
             var expectedJson = GetExpectedJson();
 
             var actualJson = JsonConvert.SerializeObject(new Net.Feature.Feature(geometry));
-
-            JsonAssert.AreEqual(expectedJson, actualJson);
-        }
-
-        [Test]
-        public void Can_Serialize_Typed_Point_Feature()
-        {
-            var geometry = new Point(new Position(1, 2));
-            var props = new TypedFeatureProps
-            {
-                Name = "no name here",
-                Value = 1.337
-            };
-            var feature = new Feature<Point, TypedFeatureProps>(geometry, props, "no id there");
-
-            var expectedJson = GetExpectedJson();
-            var actualJson = JsonConvert.SerializeObject(feature);
 
             JsonAssert.AreEqual(expectedJson, actualJson);
         }
