+++ src/GeoJSON.Net.Tests/Feature/GenericFeatureTests_Can_Deserialize_Typed_Point_Feature.json
