+++ src/GeoJSON.Net.Tests/Feature/GenericFeatureTests_Can_Serialize_Typed_Point_Feature.json
